+++ io/io_helper_functions.hpp
@@ -8,8 +8,8 @@ bool file_exists(const char *fname);
 template <class T>
 T* copyArrayAlloc(const T* source, int size);
 
-void copyArray(const std::complex<double> *source, std::complex<double> *target, long long int size);
-void copyArray(const double *source, double *target, long long int size);
+void copyArray(const std::complex<double> *source, std::complex<double> *target, long size);
+void copyArray(const double *source, double *target, long size);
 FILE * open_file(const char *fname, const char *read_write);
 
 #endif
